refactor: extract combo types to dedicated module
