+++ src/cmd/combo.rs
@@ -5,8 +5,7 @@ use tokio::{process::Command, time::Instant};
 use tracing::{debug, info, warn};
 
 use crate::{
-    ComboRunPayload, ComboRunResult, SessionEnv, SessionSocketClient, error::Result,
-    tools::RunComboOutput,
+    ComboRunPayload, ComboRunResult, RunComboOutput, SessionEnv, SessionSocketClient, error::Result,
 };
 
 pub async fn handle_combo_run(
